requirements.txt
mwclient>=0.10.1
pyyaml>=5.4.1
python-crontab>=2.6.0